net: mvpp2: limit the number of packets for TX bulk processing
The deferred update of the aggregated queue, accumulates a number
of packets to be sent in bulk. The amount directly depends on the
tx_done coalescing size. The latter is by default small, but in
case changed by ethtool to a number exceeding the aggregated queue
size, there is huge risk of a crash during forwarding test with
multiple flows. Fix the issue by limiting the possible packet
aggregation.

Signed-off-by: Marcin Wojtas <mw@semihalf.com>
Change-Id: Ib2df631e0bcdcf544dfa331e97d5d51d52440837
Reviewed-on: https://sj1git1.cavium.com/7863
Reviewed-by: Stefan Chulski <Stefan.Chulski@cavium.com>
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>