octeontx2-pf: Fix smmuv3 messages while deferring pf driver probe.
Issue occurs on probing pf driver before the af driver and the probe is
deferred with the message "AF is not initialized, deferring probe"

The reason for the issue is incorrect iova value in RVU_AF_MSIXTR_BASE as
the af driver is not probed yet.

Fixed by moving  otx2_check_pf_usable before pf tries to allocate
irq vectors.

Change-Id: I29e36556579764017c593f6914c8ca8ce02bb4e4
Signed-off-by: Christina Jacob <cjacob@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/8620
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>