arm64: defconfig: disable CONFIG_SFP
It turned out that the generic kernel SFP cage infrastructure
is incapable to work properly with all kinds of cables,
when the SFP PHY is attached indirectly to the ethernet port
in the device tree.

Example of above is MacchiatoBin Double Shot, where the cage
is a child node of the 3310 PHY. In such setup the optical SFP+
cable state is not managed appropriately by the kernel phylink
layer.

Until a proper fix is found, temporarily work-around this issue
by disabling phylink SFP handling in the kernel configuration.
As a result with all cable types will operate properly, however
using a predefined, fixed speed.

Change-Id: I1d394f5e0e592c51eeded6f6ec6c6025d7cbfc67
Signed-off-by: Marcin Wojtas <mw@semihalf.com>
Reviewed-on: https://sj1git1.cavium.com/13154
Reviewed-by: Nadav Haklai <nadavh@marvell.com>
Reviewed-by: Stefan Chulski <stefanc@marvell.com>
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>